+++ src/onegov.gazette/onegov/gazette/tests/test_forms.py
@@ -19,9 +19,13 @@ class DummyApp(object):
 
 
 class DummyRequest(object):
-    def __init__(self, session, principal=None):
+    def __init__(self, session, principal=None, private=False):
         self.app = DummyApp(session, principal)
+        self.private = private
         self.locale = 'de_CH'
+
+    def is_private(self, model):
+        return self.private
 
     def include(self, resource):
         pass
@@ -195,10 +199,10 @@ def test_notice_form(session, principal):
     assert form.validate()
 
     # Test on request
-    form = NoticeForm()
-    form.request = DummyRequest(session, principal)
-
     with freeze_time("2017-11-01 14:00"):
+        form = NoticeForm()
+        form.model = None
+        form.request = DummyRequest(session, principal)
         form.on_request()
         assert form.organization.choices == [
             ('', 'Select one'),
@@ -226,3 +230,36 @@ def test_notice_form(session, principal):
             ('13', 'Commercial Register'),
             ('14', 'Elections'),
         ]
+
+        form = NoticeForm()
+        form.model = None
+        form.request = DummyRequest(session, principal, private=True)
+        form.on_request()
+        assert form.organization.choices == [
+            ('', 'Select one'),
+            ('100', 'State Chancellery'),
+            ('200', 'Civic Community'),
+            ('300', 'Municipality'),
+            ('400', 'Evangelical Reformed Parish'),
+            ('500', 'Catholic Parish'),
+            ('600', 'Corporation')
+        ]
+        assert form.issues.choices == [
+            ('2017-44', 'No. 44, Freitag 03.11.2017'),
+            ('2017-45', 'No. 45, Freitag 10.11.2017'),
+            ('2017-46', 'No. 46, Freitag 17.11.2017'),
+            ('2017-47', 'No. 47, Freitag 24.11.2017'),
+            ('2017-48', 'No. 48, Freitag 01.12.2017'),
+            ('2017-49', 'No. 49, Freitag 08.12.2017'),
+            ('2017-50', 'No. 50, Freitag 15.12.2017'),
+            ('2017-51', 'No. 51, Freitag 22.12.2017'),
+            ('2017-52', 'No. 52, Freitag 29.12.2017'),
+            ('2018-1', 'No. 1, Freitag 05.01.2018'),
+        ]
+        assert form.category.choices == [
+            ('11', 'Education'),
+            ('12', 'Submissions'),
+            ('13', 'Commercial Register'),
+            ('14', 'Elections'),
+        ]
+        assert form.issues.render_kw['data-hot-issue'] == '2017-44'